+++ dbfread/__init__.py
@@ -15,12 +15,11 @@ __url__ = 'http://nerdly.info/ole/'
 __license__ = 'MIT'
 __version__ = '1.1.1'
 
-print(__doc__)
-
 from .dbf import DBF
 from .dbf import DBF as open  # Alias.
 from .deprecated_dbf import DeprecatedDBF as read
 
+from .exceptions import *
+
 # Prevent splat import.
 __all__ = []
-