+++ dbfread/dbf.py
@@ -12,11 +12,7 @@ from .ifiles import ifind
 from .fpt import FPT
 from .codepages import guess_encoding
 from .dbversions import get_dbversion_string
-
-try:
-    _FILE_NOT_FOUND = FileNotFoundError
-except NameError:
-    _FILE_NOT_FOUND = IOError
+from .exceptions import DataFileNotFound, MemoFileNotFound
 
 DBFHeader = StructParser(
     'DBFHeader',
@@ -118,7 +114,7 @@ class DBF(object):
         if ignorecase:
             self.filename = ifind(filename)
             if not self.filename:
-                raise _FILE_NOT_FOUND('No such file: {!r}'.format(filename))
+                raise DataFileNotFound('No such file: {!r}'.format(filename))
         else:
             self.filename = filename
 
@@ -229,7 +225,7 @@ class DBF(object):
                 self.memofilename = match
             else:
                 # Todo: warn and return field as byte string?
-                raise _FILE_NOT_FOUND('Missing memo file: {!r}'.format(fn))
+                raise MemoFileNotFound('Missing memo file: {!r}'.format(fn))
 
     def _check_headers(self):
         """Check headers for possible format errors."""