+++ dbfread/exceptions.py
@@ -1,0 +1,8 @@
+class DBFError(IOError):
+    pass
+
+class DataFileNotFound(DBFError):
+    pass
+
+class MemoFileNotFound(DBFError):
+    pass